add task_process back in
